Defaulting to OrderedDict for python 3.5 + lint all code
All code reformated by PyCharm and inspection results analysed.
