+++ eland/tests/dataframe/__init__.py
@@ -11,4 +11,3 @@
 #      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #      See the License for the specific language governing permissions and
 #      limitations under the License.
-
