+++ eland/tests/dataframe/test_repr_pytest.py
@@ -17,7 +17,7 @@
 import pandas as pd
 import pytest
 
-from eland import PY36
+from eland.compat import PY36
 from eland.dataframe import DEFAULT_NUM_ROWS_DISPLAYED
 from eland.tests.common import TestData
 
@@ -196,8 +196,8 @@ class TestDataFrameRepr(TestData):
         ed_head_str = ed_head._repr_html_()
         pd_head_str = pd_head._repr_html_()
 
-        #print(ed_head_str)
-        #print(pd_head_str)
+        # print(ed_head_str)
+        # print(pd_head_str)
 
         # Currently pandas display bold_rows=True with >=PY36 and bold_rows=False with 3.5
         # TODO - fix this test for 3.5
