+++ eland/tests/dataframe/test_to_csv_pytest.py
@@ -18,7 +18,6 @@ import ast
 import time
 
 import pandas as pd
-from elasticsearch import Elasticsearch
 from pandas.util.testing import assert_frame_equal
 
 import eland as ed
