+++ eland/tests/series/test_str_arithmetics_pytest.py
@@ -14,7 +14,6 @@
 
 # File called _pytest for PyCharm compatability
 import pytest
-import numpy as np
 
 from eland.tests.common import TestData, assert_pandas_eland_series_equal
 
@@ -60,7 +59,6 @@ class TestSeriesArithmetics(TestData):
 
         assert_pandas_eland_series_equal(pdadd, edadd)
 
-
     def test_ser_add_str_add_ser(self):
         pdadd = self.pd_ecommerce()['customer_first_name'] + self.pd_ecommerce()['customer_last_name']
         print(pdadd.name)
@@ -84,5 +82,5 @@ class TestSeriesArithmetics(TestData):
             assert self.ed_ecommerce()['customer_gender'] + self.ed_ecommerce()['customer_first_name']
 
     def test_aggregatable_add_non_aggregatable(self):
-            with pytest.raises(ValueError):
-                assert self.ed_ecommerce()['customer_first_name'] + self.ed_ecommerce()['customer_gender']
+        with pytest.raises(ValueError):
+            assert self.ed_ecommerce()['customer_first_name'] + self.ed_ecommerce()['customer_gender']
